+++ rsp.h
@@ -152,7 +152,7 @@ static short MFC0_count[32];
 
 #ifdef SP_EXECUTE_LOG
 static FILE *output_log;
-extern void step_SP_commands(unsigned long inst);
+extern void step_SP_commands(uint32_t inst);
 #endif
 extern void export_SP_memory(void);
 NOINLINE void trace_RSP_registers(void);
@@ -165,7 +165,7 @@ NOINLINE extern void run_task(void);
 #include "execute.h"
 
 #ifdef SP_EXECUTE_LOG
-void step_SP_commands(unsigned long inst)
+void step_SP_commands(uint32_t inst)
 {
     if (output_log)
     {
@@ -205,7 +205,7 @@ void step_SP_commands(unsigned long inst)
 NOINLINE void export_data_cache(void)
 {
     FILE* out;
-    register unsigned long addr;
+    register uintptr_t addr;
 
     out = fopen("rcpcache.dhex", "wb");
 #if (0)
@@ -226,7 +226,7 @@ NOINLINE void export_data_cache(void)
 NOINLINE void export_instruction_cache(void)
 {
     FILE* out;
-    register unsigned long addr;
+    register uintptr_t addr;
 
     out = fopen("rcpcache.ihex", "wb");
 #if (0)
